Naming and icon additions
