+++ src/ript_toolbar.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
 """
 /***************************************************************************
- RIPT
+ QRiS
                                  A QGIS plugin
- Riverscapes Integrated Planning Tool (RIPT)
+ QGIS Riverscapes Studio (QRiS)
  Generated by Plugin Builder: http://g-sherman.github.io/Qgis-Plugin-Builder/
                               -------------------
         begin                : 2021-05-06
@@ -32,7 +32,7 @@ from .classes.settings import Settings, CONSTANTS
 from . import resources
 
 # Import the code for the DockWidget
-from .ript_dockwidget import RIPTDockWidget
+from .qris_dockwidget import QRiSDockWidget
 from .new_project_dialog import NewProjectDialog
 from .ript_project import RiptProject
 import os.path
@@ -69,10 +69,10 @@ class RIPTToolbar:
 
         # Declare instance attributes
         self.actions = []
-        self.menu = self.tr(u'&Riverscapes Integrated Planning Tool (RIPT)')
+        self.menu = self.tr(u'&QGIS Riverscapes Studio (QRiS)')
         # TODO: We are going to let the user set this up in a future iteration
-        self.toolbar = self.iface.addToolBar(u'RIPT')
-        self.toolbar.setObjectName(u'RIPT')
+        self.toolbar = self.iface.addToolBar(u'QRiS')
+        self.toolbar.setObjectName(u'QRiS')
 
         self.settings = Settings(iface=self.iface)
 
@@ -95,7 +95,7 @@ class RIPTToolbar:
         :rtype: QString
         """
         # noinspection PyTypeChecker,PyArgumentList,PyCallByClass
-        return QCoreApplication.translate('RIPT', message)
+        return QCoreApplication.translate('QRiS', message)
 
     def add_action(
             self,
@@ -180,18 +180,18 @@ class RIPTToolbar:
         # Initialize the processing framework
         self.initProcessing()
 
-        icon_path = ':/plugins/qris_toolbar/icon.png'
+        icon_path = ':/plugins/qris_toolbar/project_qris.png'
         self.add_action(
             icon_path,
             text=self.tr(u'QRiS'),
             callback=self.run,
             parent=self.iface.mainWindow())
 
-        self.newProjectAction = QAction(QIcon(':/plugins/qris_toolbar/NewProject.png'), self.tr(u'new QRiS Project'), self.iface.mainWindow())
+        self.newProjectAction = QAction(QIcon(':/plugins/qris_toolbar/project_new_qris.png'), self.tr(u'new QRiS Project'), self.iface.mainWindow())
         self.newProjectAction.triggered.connect(self.newProjectDlg)
         self.toolbar.addAction(self.newProjectAction)
 
-        self.open_projectAction = QAction(QIcon(':/plugins/qris_toolbar/OpenProject.png'), self.tr(u'Open QRiS Project'), self.iface.mainWindow())
+        self.open_projectAction = QAction(QIcon(':/plugins/qris_toolbar/project_open_qris.png'), self.tr(u'Open QRiS Project'), self.iface.mainWindow())
         self.open_projectAction.triggered.connect(self.projectBrowserDlg)
         self.toolbar.addAction(self.open_projectAction)
 
@@ -226,7 +226,7 @@ class RIPTToolbar:
 
         for action in self.actions:
             self.iface.removePluginMenu(
-                self.tr(u'&Riverscapes Integrated Planning Tool (RIPT)'),
+                self.tr(u'&QGIS Riverscapes Studio (QRiS)'),
                 action)
             self.iface.removeToolBarIcon(action)
         # remove the toolbar
@@ -247,7 +247,7 @@ class RIPTToolbar:
             #    removed on close (see self.onClosePlugin method)
             if self.dockwidget is None:
                 # Create the dockwidget (after translation) and keep reference
-                self.dockwidget = RIPTDockWidget()
+                self.dockwidget = QRiSDockWidget()
                 self.dockwidget.iface = self.iface
 
             # connect to provide cleanup on closing of dockwidget
@@ -266,7 +266,7 @@ class RIPTToolbar:
         last_browse_path = self.settings.getValue('lastBrowsePath')
         last_dir = os.path.dirname(last_browse_path) if last_browse_path is not None else None
 
-        dialog_return = QFileDialog.getOpenFileName(self.dockwidget, "Open a RIPT project", last_dir, self.tr("RIPT Project files (project.ript)"))
+        dialog_return = QFileDialog.getOpenFileName(self.dockwidget, "Open a QRiS project", last_dir, self.tr("QRiS Project files (project.qris)"))
         if dialog_return is not None and dialog_return[0] != "" and os.path.isfile(dialog_return[0]):
             # We set the proect path in the project settings. This way it will be saved with the QgsProject file
             if self.dockwidget is None or self.dockwidget.isHidden() is True:
@@ -295,7 +295,7 @@ class RIPTToolbar:
             #    removed on close (see self.onClosePlugin method)
             if self.dockwidget is None:
                 # Create the dockwidget (after translation) and keep reference
-                self.dockwidget = RIPTDockWidget()
+                self.dockwidget = QRiSDockWidget()
                 # self.metawidget = RIPTMetaWidget()
                 # Hook metadata changes up to the metawidget
                 # self.dockwidget.metaChange.connect(self.metawidget.load)
@@ -340,11 +340,11 @@ class RIPTToolbar:
                 last_browse_path = self.settings.getValue('lastBrowsePath')
                 last_dir = os.path.dirname(last_browse_path) if last_browse_path is not None else None
 
-                dialog_return = QFileDialog.getOpenFileName(self.dockwidget, "Add GIS layer to RIPT project", last_dir, self.tr("GIS Data Sources (*.gpkg, *.tif)"))
+                dialog_return = QFileDialog.getOpenFileName(self.dockwidget, "Add GIS layer to QRiS project", last_dir, self.tr("GIS Data Sources (*.gpkg, *.tif)"))
                 if dialog_return is not None and dialog_return[0] != "" and os.path.isfile(dialog_return[0]):
                     pass
             else:
-                self.iface.messageBar().pushMessage("RIPT", "Cannot Add layer: No RIPT project is currently open.", level=1)
+                self.iface.messageBar().pushMessage("QRiS", "Cannot Add layer: No QRiS project currently open.", level=1)
 
     def addLayer(self, layer):
         pass
